+++ tests/integration/tst_WorldCommandProcessor_Regex.cpp
@@ -6,7 +6,7 @@
  * Role: QTest coverage for WorldCommandProcessor Regex behavior.
  */
 
-#include "CommandPatternUtils.h"
+#include "WorldCommandProcessorUtils.h"
 
 #include <QtTest/QTest>
 
@@ -77,7 +77,6 @@ class tst_WorldCommandProcessor_Regex : public QObject
 };
 
 QTEST_APPLESS_MAIN(tst_WorldCommandProcessor_Regex)
-
 
 #if __has_include("tst_WorldCommandProcessor_Regex.moc")
 #include "tst_WorldCommandProcessor_Regex.moc"
